+++ src/main/java/com/github/dakusui/crest/Crest.java
@@ -2,10 +2,11 @@ package com.github.dakusui.crest;
 
 import com.github.dakusui.crest.core.*;
 import com.github.dakusui.crest.core.Call.Arg;
+import com.github.dakusui.crest.functions.Executable;
 import com.github.dakusui.crest.matcherbuilders.*;
 import com.github.dakusui.crest.matcherbuilders.primitives.*;
-import com.github.dakusui.crest.utils.printable.Functions;
-import com.github.dakusui.crest.utils.printable.Printable;
+import com.github.dakusui.crest.functions.printable.Functions;
+import com.github.dakusui.crest.functions.printable.Printable;
 import org.opentest4j.AssertionFailedError;
 import org.opentest4j.TestAbortedException;
 
@@ -15,9 +16,9 @@ import java.util.function.Function;
 import java.util.function.Predicate;
 
 import static com.github.dakusui.crest.utils.InternalUtils.*;
-import static com.github.dakusui.crest.utils.printable.Functions.trivial;
-import static com.github.dakusui.crest.utils.printable.Predicates.equalTo;
-import static com.github.dakusui.crest.utils.printable.Predicates.isEmptyArray;
+import static com.github.dakusui.crest.functions.printable.Functions.trivial;
+import static com.github.dakusui.crest.functions.printable.Predicates.equalTo;
+import static com.github.dakusui.crest.functions.printable.Predicates.isEmptyArray;
 import static java.util.Arrays.asList;
 import static java.util.Objects.requireNonNull;
