+++ src/main/java/com/github/dakusui/crest/core/Session.java
@@ -1,7 +1,7 @@
 package com.github.dakusui.crest.core;
 
 import com.github.dakusui.crest.functions.TransformingPredicate;
-import com.github.dakusui.crest.utils.printable.TrivialFunction;
+import com.github.dakusui.crest.functions.printable.TrivialFunction;
 import org.opentest4j.AssertionFailedError;
 
 import java.util.*;
