+++ src/main/java/com/github/dakusui/crest/functions/printable/Printable.java
@@ -1,4 +1,4 @@
-package com.github.dakusui.crest.utils.printable;
+package com.github.dakusui.crest.functions.printable;
 
 
 import java.util.function.Function;
