+++ src/main/java/com/github/dakusui/crest/functions/printable/PrintableFunction.java
@@ -1,4 +1,4 @@
-package com.github.dakusui.crest.utils.printable;
+package com.github.dakusui.crest.functions.printable;
 
 import java.util.Objects;
 import java.util.function.Function;
