+++ src/main/java/com/github/dakusui/crest/matcherbuilders/MatcherBuilder.java
@@ -1,7 +1,7 @@
 package com.github.dakusui.crest.matcherbuilders;
 
 import com.github.dakusui.crest.core.Matcher;
-import com.github.dakusui.crest.utils.printable.Predicates;
+import com.github.dakusui.crest.functions.printable.Predicates;
 
 import java.util.List;
 import java.util.function.Function;
