+++ src/test/java/com/github/dakusui/crest/examples/InThincrest.java
@@ -1,7 +1,7 @@
 package com.github.dakusui.crest.examples;
 
 import com.github.dakusui.crest.Crest;
-import com.github.dakusui.crest.utils.printable.Functions;
+import com.github.dakusui.crest.functions.printable.Functions;
 import org.hamcrest.CoreMatchers;
 import org.hamcrest.core.StringContains;
 import org.junit.Test;
@@ -9,9 +9,9 @@ import org.junit.Test;
 import java.util.*;
 
 import static com.github.dakusui.crest.Crest.*;
-import static com.github.dakusui.crest.utils.printable.Functions.*;
-import static com.github.dakusui.crest.utils.printable.Predicates.eq;
-import static com.github.dakusui.crest.utils.printable.Predicates.isEmpty;
+import static com.github.dakusui.crest.functions.printable.Functions.*;
+import static com.github.dakusui.crest.functions.printable.Predicates.eq;
+import static com.github.dakusui.crest.functions.printable.Predicates.isEmpty;
 
 /**
  * http://qiita.com/disc99/items/31fa7abb724f63602dc9
