+++ src/test/java/com/github/dakusui/crest/ut/MatcherTest.java
@@ -1,8 +1,8 @@
 package com.github.dakusui.crest.ut;
 
 import com.github.dakusui.crest.core.Matcher;
-import com.github.dakusui.crest.utils.printable.Functions;
-import com.github.dakusui.crest.utils.printable.Predicates;
+import com.github.dakusui.crest.functions.printable.Functions;
+import com.github.dakusui.crest.functions.printable.Predicates;
 import org.junit.Test;
 import org.opentest4j.AssertionFailedError;
