+++ src/test/java/com/github/dakusui/crest/utils/ut/PredicatesTest.java
@@ -1,6 +1,6 @@
 package com.github.dakusui.crest.utils.ut;
 
-import com.github.dakusui.crest.utils.printable.Predicates;
+import com.github.dakusui.crest.functions.printable.Predicates;
 import org.junit.Test;
 import org.junit.experimental.runners.Enclosed;
 import org.junit.runner.RunWith;
